requirements.txt
pandas
streamlit==0.72.0
pytorch-lightning==1.1.0
openpyxl